some tweaks (#851)
