+++ util/mod.ts
@@ -10,4 +10,5 @@ export * from "./memo.ts"
 export * from "./normalize.ts"
 export * from "./notifier.ts"
 export * from "./state.ts"
+export * from "./waitFor.ts"
 export * from "./withSignal.ts"
